gcp/pam/cloud-functions/pam-request-grant-approve/requirements.txt
cachetools 5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
deprecation==2.1.0
flask==2.2.5
functions-framework==3.8.1
gunicorn==23.0.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-secret-manager==2.22.0
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio-status==1.68.1
grpcio==1.68.1
gunicorn==23.0.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
markupsafe==2.1.5
packaging==24.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1-modules==0.4.1
pyasn1==0.6.1
requests==2.32.3
rsa==4.9
typing-extensions==4.7.1
urllib3==2.2.3
watchdog==3.0.0
werkzeug==2.2.3
zipp==3.15.0
